chore: Extract PointerCapture dedicated files
