+++ src/Uno.UI/UI/Xaml/Internal/PointerCaptureKind.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Linq;
+
+namespace Uno.UI.Xaml.Core;
+
+[Flags]
+internal enum PointerCaptureKind : byte
+{
+	None = 0,
+
+	Explicit = 1,
+	Implicit = 2,
+
+	Any = Explicit | Implicit,
+}
