+++ src/Uno.UI/UI/Xaml/UIElement.Pointers.Managed.cs
@@ -104,10 +104,10 @@ namespace Windows.UI.Xaml
 
 		#endregion
 
-		partial void CapturePointerNative(Pointer pointer)
+		internal partial void CapturePointerNative(Pointer pointer)
 			=> XamlRoot?.VisualTree.ContentRoot.InputManager!.SetPointerCapture(pointer.UniqueId);
 
-		partial void ReleasePointerNative(Pointer pointer)
+		internal partial void ReleasePointerNative(Pointer pointer)
 			=> XamlRoot?.VisualTree.ContentRoot.InputManager!.ReleasePointerCapture(pointer.UniqueId);
 	}
 }
